backend/requirements.txt
Flask>=2.0.0
gunicorn==20.1.0
kafka-python==2.0.2
psycopg2-binary==2.9.1
flask_sqlalchemy
